+++ services/http/middleware/PrepareAppInfo.js
@@ -1,6 +1,10 @@
 const AbstractMiddleware = require('./AbstractMiddleware');
 
 class PrepareAppInfo extends AbstractMiddleware {
+  static get description() {
+    return 'Basic middleware that creates "req.appInfo" object';
+  }
+
   async middleware(req, res, next) {
     if (!req.appInfo) {
       req.appInfo = {