+++ tests/setup.js
@@ -9,7 +9,7 @@ const Server = require('../server');
 jest.setTimeout(1000000);
 beforeAll(async () => {
   mongoMemoryServerInstance = await MongoMemoryReplSet.create({
-    binary: { version: '4.4.6' },
+    // binary: { version: '4.4.6' },
     replSet: { storageEngine: 'wiredTiger' },
   });
   await mongoMemoryServerInstance.waitUntilRunning();